+++ cmd/client/main.go
@@ -90,12 +90,13 @@ func main() {
 		},
 		Commands: []*cli.Command{
 			initCmd,
-			initDidCmd,
-			recoverCmd,
+			recoverClietnCmd,
 			netCmd,
 			modelCmd,
 			fileCmd,
 			didCmd,
+			reportFaultsCmd,
+			recoverFaultsCheckRequestCmd,
 			account.AccountCmd,
 			cliutil.GenerateDocCmd,
 		},
@@ -155,7 +156,7 @@ var initCmd = &cli.Command{
 				fmt.Printf("%v", err)
 				askFor = true
 			} else {
-				if coins.AmountOf("sao").LT(math.NewInt(1000)) {
+				if coins.AmountOf(chain.DENOM).LT(math.NewInt(1000)) {
 					askFor = true
 				}
 			}
@@ -194,88 +195,8 @@ var initCmd = &cli.Command{
 	},
 }
 
-var initDidCmd = &cli.Command{
-	Name:  "initDid",
-	Usage: "initialize a Decentralized Identifier (DID) on the sao client",
-	UsageText: "To use sao CLI with Decentralized Identifiers (DID), first initialize it using this command.\n " +
-		"This creates a sao chain account locally that will be used as default account in subsequent commands. \n" +
-		"In the --repo directory, you will find the client configuration file.\n" +
-		"In the --keyring directory, you will find the keystore files.",
-	Flags: []cli.Flag{
-		&cli.StringFlag{
-			Name:     cliutil.FlagKeyName,
-			Usage:    "sao chain account key name",
-			Required: true,
-			Aliases:  []string{"k"},
-		},
-		&cli.StringFlag{
-			Name:     "chain-id",
-			Required: false,
-			Value:    "sao",
-		},
-		&cli.StringFlag{
-			Name:     "address",
-			Required: false,
-			Value:    "",
-		},
-	},
-	Action: func(cctx *cli.Context) error {
-		repo := cctx.String(FlagClientRepo)
-
-		saoclient, closer, err := getSaoClient(cctx)
-		if err != nil {
-			return err
-		}
-		defer closer()
-		fmt.Printf("repo %s is initialized.", repo)
-		fmt.Println()
-
-		address := cctx.String("address")
-		fmt.Println("account created: ")
-		fmt.Println("Address:", address)
-
-		for {
-			coins, err := saoclient.GetBalance(cctx.Context, address)
-			askFor := false
-			if err != nil {
-				fmt.Printf("%v", err)
-				askFor = true
-			} else {
-				if coins.AmountOf("sao").LT(math.NewInt(1000)) {
-					askFor = true
-				}
-			}
-			if askFor {
-				fmt.Print("Please deposit enough coins to pay gas. Confirm with 'yes' :")
-				reader := bufio.NewReader(os.Stdin)
-				indata, err := reader.ReadBytes('\n')
-				if err != nil {
-					return err
-				}
-				_ = strings.Replace(string(indata), "\n", "", -1)
-			} else {
-				break
-			}
-		}
-
-		didManager, address, err := cliutil.GetDidManager(cctx, saoclient.Cfg.KeyName)
-		if err != nil {
-			return err
-		}
-		fmt.Println("DID:", didManager.Id)
-
-		hash, err := saoclient.UpdateDidBinding(cctx.Context, address, didManager.Id, fmt.Sprintf("cosmos:%s:%s", cctx.String("chain-id"), address))
-		if err != nil {
-			return err
-		}
-
-		fmt.Printf("Created DID %s. tx hash %s", didManager.Id, hash)
-		return nil
-	},
-}
-
-var recoverCmd = &cli.Command{
-	Name:  "recover",
+var recoverClietnCmd = &cli.Command{
+	Name:  "recover-cli",
 	Usage: "recover cli sao client with a specific did",
 	UsageText: "if you have already init sao cli client, you can do recover client by did.\n " +
 		"return error if did is not exists or payment address of did is not found in keyring directory ",
@@ -333,6 +254,98 @@ var recoverCmd = &cli.Command{
 
 		fmt.Println()
 		fmt.Println("sao client recovered.")
+		return nil
+	},
+}
+
+var reportFaultsCmd = &cli.Command{
+	Name:  "report-faults",
+	Usage: "report the storage faults",
+	Flags: []cli.Flag{
+		&cli.StringSliceFlag{
+			Name:     "data-ids",
+			Usage:    "faulted data model id list to report",
+			Required: true,
+		},
+	},
+	Action: func(cctx *cli.Context) error {
+		ctx := cctx.Context
+
+		client, closer, err := getSaoClient(cctx)
+		if err != nil {
+			return err
+		}
+		defer closer()
+
+		dataIds := cctx.StringSlice("data-ids")
+		if len(dataIds) == 0 {
+			return types.Wrapf(types.ErrInvalidParameters, "data-ids is required")
+		}
+
+		if tx, err := client.FaultsCheck(ctx, dataIds); err != nil {
+			return err
+		} else {
+			fmt.Println(tx)
+		}
+
+		return nil
+	},
+}
+
+var recoverFaultsCheckRequestCmd = &cli.Command{
+	Name:  "requst-faults-recover-check",
+	Usage: "requst the fishmen to check the recoverable storage faults",
+	Flags: []cli.Flag{
+		&cli.StringFlag{
+			Name:     "provider",
+			Usage:    "provider with recoverable storage faults",
+			Required: true,
+		},
+		&cli.StringSliceFlag{
+			Name:     "fault-ids",
+			Usage:    "faults id list to requst",
+			Required: true,
+		},
+	},
+	Action: func(cctx *cli.Context) error {
+		ctx := cctx.Context
+
+		client, closer, err := getSaoClient(cctx)
+		if err != nil {
+			return err
+		}
+		defer closer()
+
+		creator, err := client.GetNodeAddress(ctx)
+		if err != nil {
+			return err
+		}
+
+		fishmen, err := client.GetFishmen(ctx)
+		if err != nil {
+			return err
+		}
+
+		if !strings.Contains(fishmen, creator) {
+			return types.Wrapf(types.ErrInvalidParameters, "curren gateway is not a fishmen")
+		}
+
+		provider := cctx.String("provider")
+		if provider == "" {
+			return types.Wrapf(types.ErrInvalidParameters, "provider is required")
+		}
+
+		faultIds := cctx.StringSlice("fault-ids")
+		if len(faultIds) == 0 {
+			return types.Wrapf(types.ErrInvalidParameters, "fault-ids is required")
+		}
+
+		if tx, err := client.RecoverCheck(ctx, provider, faultIds); err != nil {
+			return err
+		} else {
+			fmt.Println(tx)
+		}
+
 		return nil
 	},
 }