+++ CHANGELOG.md
@@ -1,8 +1,16 @@
 <a name="unreleased"></a>
 ## [Unreleased]
 
+### Features
+- update shard job to expire status after expiration
+- add storage command ([#21](https://github.com/SAONetwork/sao-node.git/issues/21))
+- add rpc to load model delegated by gateway ([#20](https://github.com/SAONetwork/sao-node.git/issues/20))
+- poe ([#17](https://github.com/SAONetwork/sao-node.git/issues/17))
+- remove expire shard from storage ([#19](https://github.com/SAONetwork/sao-node.git/issues/19))
+
 ### Bug Fixes
-- try restore timeout map when node start
+- meta status ([#22](https://github.com/SAONetwork/sao-node.git/issues/22))
+- add startup pledge check ([#18](https://github.com/SAONetwork/sao-node.git/issues/18))
 - fetch replica mode content, handle invalid cid, model cache evict ([#13](https://github.com/SAONetwork/sao-node.git/issues/13))